hero animation and better ui
