+++ flutter_application_1/lib/Pages/home_page.dart
@@ -2,15 +2,16 @@
 // ignore_for_file: prefer_const_constructors
 import 'dart:convert';
 //import 'dart:html';
-
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-
 import 'package:flutter_application_1/models/catalog.dart';
 import 'package:flutter_application_1/widgets/themes.dart';
 import 'package:flutter_application_1/widgets/drawer.dart';
-
 import 'package:velocity_x/velocity_x.dart';
+
+import '../widgets/home_widgets/CatalogHeader.dart';
+import '../widgets/home_widgets/CatalogList.dart';
+
 //import 'package:flutter_application_1/widgets/item_widget.dart';
 
 
@@ -59,11 +60,10 @@ class _HomepageState extends State<Homepage> {
             children: [
                CatalogHeader(),
                if(CatalogModel.items!=null && CatalogModel.items!.isNotEmpty)
-                CatalogList().expand()
+                CatalogList().py32().expand() //to take full screen
                else 
-                Center(
-                  child: CircularProgressIndicator(),
-                ) 
+                  CircularProgressIndicator().centered().expand(),
+                 
                 
       
             ]
@@ -138,107 +138,5 @@ class _HomepageState extends State<Homepage> {
 //       ),
 //      drawer: MyDrawer(),
     );
-  }
-}
-
-class CatalogHeader extends StatelessWidget {
-  const CatalogHeader({ Key? key }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-            crossAxisAlignment: CrossAxisAlignment.start, // to make the text from left
-            // ignore: prefer_const_literals_to_create_immutables
-            children: [
-               "Catalog App".text.xl5.bold.color(MyTheme.DarkBluishCOlor).make(),
-               "Trending products".text.xl2.make(),
-      
-            ]
-    );
-  }
-}
-
-class CatalogList extends StatelessWidget {
-  const CatalogList({ Key? key }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return ListView.builder(
-      shrinkWrap: true, //to stop listview take full screen
-      itemCount: CatalogModel.items?.length ,
-      itemBuilder: (context,index){
-        final catalog= CatalogModel.items![index];
-        return CatalogItem(catalog:catalog);
-      },
-      
-    );
-  }
-}
-
-class CatalogItem extends StatelessWidget {
-  const CatalogItem({
-    Key? key,
-    required this.catalog,
-  }) : assert(catalog!=null),
-   super(key: key);
-  final Item catalog;
-  @override
-  Widget build(BuildContext context) {
-    return VxBox(
-      child: Row(
-        children: [
-         CatalogImage(
-            image: catalog.image,
-          ),
-          Expanded(child:  Column(
-            crossAxisAlignment: CrossAxisAlignment.start, // to start text from left  <------> left right
-            mainAxisAlignment: MainAxisAlignment.center,// up-down alignment
-            children: [
-              catalog.name.text.lg.bold.color(MyTheme.DarkBluishCOlor).make(),
-              catalog.desc.text.textStyle(context.captionStyle).make(),
-              10.heightBox,
-              ButtonBar(
-                alignment: MainAxisAlignment.spaceBetween, //to get max dist between two horizontal thing
-                //buttonPadding: Vx.mOnly(right: 16),
-                buttonPadding: EdgeInsets.zero,
-                children: [
-                  "\$${catalog.price}".text.bold.xl.make(),
-                  ElevatedButton(
-                    onPressed: () {}, 
-                    style: ButtonStyle(
-                      backgroundColor: MaterialStateProperty.all(
-                        MyTheme.DarkBluishCOlor,
-                      ),
-
-                      shape: MaterialStateProperty.all(
-                        StadiumBorder(), 
-                      )
-                    ),
-                    child: "Buy".text.make()
-                  )
-                ],
-              ).pOnly(right: 8.0),
-            ],
-            )
-          )
-        ],
-      ),
-    ).white.roundedLg.square(150).make().p16();
-  }
-}
-class CatalogImage extends StatelessWidget {
-  final String image;
-  const CatalogImage({
-    Key? key,
-    required this.image,
-  }) :super(key: key);
-  
-  @override
-  Widget build(BuildContext context) {
-    return  Image.network(
-            //catalog.image
-            image,
-          ).box.color(MyTheme.creamColor).make().p16().w32(context);
-          //w is width, p is padding
   }
 }
