+++ flutter_application_1/lib/utils/routes.dart
@@ -1,5 +1,6 @@
 class MyRoutes{
   static String loginRoute="/login";
   static String homeRoute="/home";
+  static String homeDetailsRoute = "/detail";
   
 }
