+++ flutter_application_1/lib/widgets/home_widgets/CatalogHeader.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_application_1/widgets/themes.dart';
+import 'package:velocity_x/velocity_x.dart';
+
+class CatalogHeader extends StatelessWidget {
+  const CatalogHeader({ Key? key }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+            crossAxisAlignment: CrossAxisAlignment.start, // to make the text from left
+            // ignore: prefer_const_literals_to_create_immutables
+            children: [
+               "Catalog App".text.xl5.bold.color(MyTheme.DarkBluishCOlor).make(),
+               "Trending products".text.xl2.make(),
+      
+            ]
+    );
+  }
+}
